Fix T48977: rigged meshes' animation is not really supported in FBX.
rigged meshes are global, but implicitly local to their armature... so far
was exporting their animation in global space, and trying to correct it back to local
on import, but this cannot work that way - and actually does not make much sense.

So for now, fully disabling tranform animation of rigged meshes. Did quick check with
Unreal engine, and looks like it is totally ignoring any rigged mesh animation anyway.

If a kind miracle gives us some day full specs of that format, we may learn
what exact behavior is expected in that case.